Fix merge conflicts and Reducers Folder
